requirements.txt
PyQt6>=6.10.0
python-can>=4.6.0
cantools>=41.0.0
pyserial>=3.5
rich>=14.2.0
